Add an asynchronous reading state machine for TCP clients
This state machine defines a set of pure virtual methods to be called
depending on the current state machine state.
The available states are:
1. Idle: This is the initial state of the client.
2. Authing: This state is the client first active state. In this state,
the client has to provide vaild login credintials in order to start
reading and sending data.
3. Reading Header: In this state the client reads the next message
header which describes the message details.
4. Reading Body: In this state the client is reading the actual message
contents.
5. Closing: In this state, the client is on destruction phase. This
state is entered when requested by the user, a privileged user, or on
errors.