Add a queue-based send method to the tcp_client class
Messages queue is a vector of string shared pointers. The call has to
enqueue the messages before sending them using the method `enqueue_send'
which accepts std::shared_ptr<std::string> as a parameter. Then, the
call has to call the method `send_enqueued' to send all messages in the
queue in the right order.